Time: 11 ms (7.37%), Space: 9.2 MB (5.01%) - LeetHub

+++ 0790-domino-and-tromino-tiling/0790-domino-and-tromino-tiling.cpp
@@ -27,6 +27,6 @@ private:
     
 public:
     int numTilings(int n) {
-        return f(n);
+        return static_cast<int>(f(n));
     }
 };
